pyproject.toml
[project]
name = "ai-inetrnships-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
]
